backend/requirements.txt
Django
djangorestframework
django-filter
mysqlclient
pymysql
djangorestframework-simplejwt
django-cors-headers
python-dotenv
celery==5.6.3
redis==7.4.0
gunicorn
cryptography
whitenoise